Move tests of .transport.files to parallel module
